Reuse delete modal for sending empty messages
